Remove common.py and put common checks into inheritable class in checks.py
Add arguments to super() calls
